file naming changed:see detailed commit message 	renamed:    src/controllers/AdminControllers.mjs -> controllers/admin.controller.mjs 	renamed:    src/controllers/cartControllers.mjs -> controllers/cart.controller.mjs 	renamed:    src/controllers/CouponControllers.mjs -> controllers/coupon.controller.mjs 	renamed:    src/controllers/FavouriteControllers.mjs -> controllers/favourite.controller.mjs 	modified:   controllers/index.mjs 	renamed:    src/controllers/TestimonialControllers.mjs -> controllers/order.controller.mjs 	renamed:    src/controllers/orderControllers.mjs -> controllers/payment.controller.mjs 	renamed:    src/controllers/productControllers.mjs -> controllers/product.controller.mjs 	renamed:    src/controllers/paymentControllers.mjs -> controllers/promotion.controller.mjs 	renamed:    src/controllers/promotionControllers.mjs -> controllers/review.controller.mjs 	renamed:    src/controllers/userControllers.mjs -> controllers/testimonial.controller.mjs 	renamed:    src/models/AdminModels.mjs -> models/admin.model.mjs 	renamed:    src/models/cartModels.mjs -> models/cart.model.mjs 	renamed:    src/models/chatModel.mjs -> models/chat.model.mjs 	renamed:    src/models/CouponModels.mjs -> models/coupon.model.mjs 	renamed:    src/models/FavouriteModels.mjs -> models/favourite.model.mjs 	renamed:    src/models/orderModels.mjs -> models/order.model.mjs 	renamed:    src/models/paymentModels.mjs -> models/payment.model.mjs 	renamed:    src/models/productModels.mjs -> models/product.model.mjs 	renamed:    src/models/promotionModels.mjs -> models/promotion.model.mjs 	renamed:    src/models/reviewModels.mjs -> models/review.model.mjs 	renamed:    src/models/salesinfoModel.mjs -> models/salesinfo.model.mjs 	renamed:    src/models/sessioninfoModel.mjs -> models/sessioninfo.model.mjs 	renamed:    src/models/testimonialModels (2).mjs -> models/testimonial.model.mjs 	modified:   models/user.model.mjs 	renamed:    src/routes/AdminRoutes.js -> routes/v1/admin.route.mjs 	renamed:    src/routes/CouponRoutes.mjs -> routes/v1/cart.route.mjs 	renamed:    src/routes/FavouriteRoutes.mjs -> routes/v1/coupon.route.mjs 	renamed:    src/routes/TestimonialRoutes.mjs -> routes/v1/favourite.route.mjs 	modified:   routes/v1/index.mjs 	renamed:    src/routes/cartRoutes.mjs -> routes/v1/order.route.mjs 	renamed:    src/routes/orderRoutes.mjs -> routes/v1/payment.route.mjs 	renamed:    src/routes/productRoutes.mjs -> routes/v1/product.route.mjs 	renamed:    src/routes/paymentRoutes.mjs -> routes/v1/promotion.route.mjs 	renamed:    src/routes/promotionRoutes.mjs -> routes/v1/review.route.mjs 	renamed:    src/routes/userRoutes.mjs -> routes/v1/testimonial.route.mjs 	deleted:    src/controllers/index.mjs 	deleted:    src/models/userModels.mjs
