+++ controllers/index.mjs
@@ -1,2 +1,32 @@
-module.exports.authController = require('./auth.controller');
-module.exports.userController = require('./user.controller');
+// controllers/index.mjs
+
+// Import all controllers
+import * as authController from './auth.controller.mjs';
+import * as userController from './user.controller.mjs';
+import * as adminController from './admin.controller.mjs';
+import * as cartController from './cart.controller.mjs';
+import * as couponController from './coupon.controller.mjs';
+import * as favouriteController from './favourite.controller.mjs';
+import * as orderController from './order.controller.mjs';
+import * as paymentController from './payment.controller.mjs';
+import * as productController from './product.controller.mjs';
+import * as promotionController from './promotion.controller.mjs';
+import * as reviewController from './review.controller.mjs';
+import * as testimonialController from './testimonial.controller.mjs';
+
+// Export all controllers
+export {
+  authController,
+  userController,
+  adminController,
+  cartController,
+  couponController,
+  favouriteController,
+  orderController,
+  paymentController,
+  productController,
+  promotionController,
+  reviewController,
+  testimonialController,
+};
+
