+++ routes/v1/index.mjs
@@ -1,6 +1,18 @@
+// routes/index.mjs
+
 import express from 'express';
 import authRoute from './auth.route.mjs';
 import userRoute from './user.route.mjs';
+// import adminRoute from './admin.route.mjs';
+// import cartRoute from './cart.route.mjs';
+// import couponRoute from './coupon.route.mjs';
+// import favouriteRoute from './favourite.route.mjs';
+// import orderRoute from './order.route.mjs';
+// import paymentRoute from './payment.route.mjs';
+// import productRoute from './product.route.mjs';
+// import promotionRoute from './promotion.route.mjs';
+// import reviewRoute from './review.route.mjs';
+// import testimonialRoute from './testimonial.route.mjs';
 import docsRoute from './docs.route.mjs';
 import config from '../../config/config.mjs';
 
@@ -15,6 +27,46 @@ const defaultRoutes = [
     path: '/users',
     route: userRoute,
   },
+  // {
+  //   path: '/admin',
+  //   route: adminRoute,
+  // },
+  // {
+  //   path: '/cart',
+  //   route: cartRoute,
+  // },
+  // {
+  //   path: '/coupon',
+  //   route: couponRoute,
+  // },
+  // {
+  //   path: '/favourite',
+  //   route: favouriteRoute,
+  // },
+  // {
+  //   path: '/order',
+  //   route: orderRoute,
+  // },
+  // {
+  //   path: '/payment',
+  //   route: paymentRoute,
+  // },
+  // {
+  //   path: '/product',
+  //   route: productRoute,
+  // },
+  // {
+  //   path: '/promotion',
+  //   route: promotionRoute,
+  // },
+  // {
+  //   path: '/review',
+  //   route: reviewRoute,
+  // },
+  // {
+  //   path: '/testimonial',
+  //   route: testimonialRoute,
+  // },
 ];
 
 const devRoutes = [
@@ -37,4 +89,3 @@ if (config.env === 'development') {
 }
 
 export default router;
-
